move binning for plot into utils.pandas
