+++ moonstone/utils/pandas/series.py
@@ -1,7 +1,14 @@
+import math
+import numpy as np
+import pandas as pd
+
+from moonstone.utils.plot import (
+    check_list_type
+)
 from moonstone.utils.convert import pandas_to_python_type
 
 
-class SeriesStatsBuilder(object):
+class SeriesStatsBuilder:
 
     def __init__(self, series):
         self.series = series
@@ -34,3 +41,54 @@ class SeriesStatsBuilder(object):
     def build_stats(self):
         return getattr(self, f"_build_{str(self.series.dtype)}_stats",
                        self._build_base_stats)()
+
+
+class SeriesBinning:
+
+    def __init__(self, series: pd.Series):
+        self.data = series
+
+    def compute_heterogeneous_bins(self):
+        """Logish bins"""
+        maximum = self.data.max()
+        magnitude = int(math.log10(maximum))
+        bval = [-0.1]  # to have the 0 value
+        i = 0
+        while i < magnitude:
+            bval += list(np.arange(2*10**i, 10**(i+1)+1, 10**i))
+            i += 1
+        # i=magnitude
+        bval += list(np.arange(2*10**i, maximum+10**i, 10**i))  # up to maximum
+        return bval
+
+    @property
+    def bins_values(self):
+        """
+        retrieves bins_values, and compute it if no values given
+        """
+        if getattr(self, '_bins_values', None) is None:
+            self.bins_values = self.compute_heterogeneous_bins()
+        return self._bins_values
+
+    @bins_values.setter
+    def bins_values(self, bins_values):
+        if type(bins_values) == list and check_list_type(bins_values, (int, float, np.integer)):
+            self._bins_values = bins_values
+        else:
+            raise ValueError("Error : expecting list of numerical values (int, float) in bins_values.")
+
+    def compute_binned_data(self, normalize: bool = False):
+        binned_df = pd.cut(self.data, bins=self.bins_values)   # put every items in the appropriate bin
+        data = pd.value_counts(binned_df, normalize=normalize)
+        data = data.reindex(binned_df.cat.categories)
+        new_xnames = list(data.index.astype(str))
+        new_xnames[0] = new_xnames[0].replace('(-0.1', '[0.0')
+        new_xnames = [new_xnames[i].replace('(', ']') for i in range(len(new_xnames))]
+        data.index = new_xnames
+        return data
+
+    @property
+    def binned_data(self):
+        if getattr(self, '_binned_data', None) is None:
+            self._binned_data = self.compute_binned_data()
+        return self._binned_data
